code refactor: rename ellar.common.logger to ellar.common.logging
