+++ ellar/common/params/args/request_model.py
@@ -2,7 +2,7 @@ import typing as t
 
 from ellar.common.constants import MULTI_RESOLVER_KEY
 from ellar.common.interfaces import IExecutionContext
-from ellar.common.logger import logger
+from ellar.common.logging import logger
 from ellar.pydantic import (
     BaseModel,
     ModelField,
