+++ ellar/common/params/resolvers/parameter.py
@@ -11,7 +11,7 @@ from ellar.common.constants import (
 )
 from ellar.common.exceptions import RequestValidationError
 from ellar.common.interfaces import IExecutionContext
-from ellar.common.logger import request_logger
+from ellar.common.logging import request_logger
 from ellar.pydantic import (
     is_sequence_field,
     lenient_issubclass,
