+++ ellar/common/responses/models/json.py
@@ -2,7 +2,7 @@ import typing as t
 
 from ellar.common.constants import SERIALIZER_FILTER_KEY
 from ellar.common.interfaces import IExecutionContext
-from ellar.common.logger import request_logger
+from ellar.common.logging import request_logger
 from ellar.common.serializer import SerializerFilter, serialize_object
 from ellar.pydantic import as_pydantic_validator, create_model_field
 from ellar.reflect import reflect
